[Refactor] Endpoint 구조체 및 API 모델 구조체 디렉토리 수정 (#41)
