--- BookKitty/BookKitty/BookMatchKit/Sources/BookMatchAPI/Models/DTOs/ChatMessage.swift
@@ -1,4 +1,0 @@
-struct ChatMessage: Codable {
-    let role: String
-    let content: String
-}
