+++ BookKitty/BookKitty/BookMatchKit/Sources/BookMatchAPI/Models/Endpoints/ImageDownloadEndpoint.swift
@@ -2,10 +2,6 @@ import Foundation
 import Network
 
 struct ImageDownloadEndpoint: Endpoint {
-    // MARK: Lifecycle
-
-    // MARK: Initialization
-
     init(urlString: String) {
         self.urlString = urlString
     }
@@ -15,16 +11,16 @@ struct ImageDownloadEndpoint: Endpoint {
     typealias Response = Data
 
     // MARK: Endpoint Protocol
-
+    /// Cannot use instance member 'urlString' within property initializer; property initializers run before 'self' is available
     var baseURL: String {
         urlString
     }
 
-    var path: String { "" }
+    var path = ""
 
-    var method: HTTPMethod { .get }
+    var method = HTTPMethod.get
 
-    var heaerFields: [String: String] { [:] }
+    var headerFields: [String: String] { [:] }
 
     var queryItems: [URLQueryItem] { [] }
