+++ BookKitty/BookKitty/BookMatchKit/Sources/BookMatchAPI/Models/Endpoints/NaverBooksEndpoint.swift
@@ -2,10 +2,6 @@ import Foundation
 import Network
 
 struct NaverBooksEndpoint: Endpoint {
-    // MARK: Lifecycle
-
-    // MARK: - Initialization
-
     init(query: String, limit: Int, configuration: APIConfiguration) {
         self.query = query
         self.limit = limit
@@ -16,19 +12,11 @@ struct NaverBooksEndpoint: Endpoint {
 
     typealias Response = NaverBooksResponse
 
-    let query: String
-    let limit: Int
-    let configuration: APIConfiguration
+    var baseURL = "https://openapi.naver.com"
+    var path = "/v1/search/book.json"
+    var method = HTTPMethod.get
 
-    // MARK: - Endpoint Protocol
-
-    var baseURL: String { "https://openapi.naver.com" }
-
-    var path: String { "/v1/search/book.json" } // baseURL에 이미 전체 경로가 포함되어 있으므로 빈 문자열
-
-    var method: HTTPMethod { .get }
-
-    var heaerFields: [String: String] {
+    var headerFields: [String: String] {
         [
             "X-Naver-Client-Id": configuration.naverClientId,
             "X-Naver-Client-Secret": configuration.naverClientSecret,
@@ -45,5 +33,9 @@ struct NaverBooksEndpoint: Endpoint {
 
     var timeoutInterval: TimeInterval { 30.0 } // 기본 30초 타임아웃
 
-    var data: Data? { nil } // GET 요청이므로 body 데이터 없음
+    var data: Data? { nil }
+    
+    private let query: String
+    private let limit: Int
+    private let configuration: APIConfiguration
 }
