+++ BookKitty/BookKitty/BookMatchKit/Sources/BookMatchCore/Models/BookMatchModuleOutput.swift
@@ -1,21 +1,3 @@
-/// 도서 매칭 모듈의 입력 데이터를 나타내는 구조체입니다.
-public struct BookMatchModuleInput {
-    // MARK: Lifecycle
-
-    public init(
-        question: String,
-        ownedBooks: [OwnedBook]
-    ) {
-        self.question = question
-        self.ownedBooks = ownedBooks
-    }
-
-    // MARK: Public
-
-    public let question: String
-    public let ownedBooks: [OwnedBook]
-}
-
 /// 도서 매칭 모듈의 출력 데이터를 나타내는 구조체입니다.
 public struct BookMatchModuleOutput {
     // MARK: Lifecycle
